FitLink-Backend/pyproject.toml
[tool.poetry]
name = "fitlink-backend"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["Valeriacarmonaf <97414518+Valeriacarmonaf@users.noreply.github.com>"]
packages = [{ include = "fitlink_backend", from = "src" }]

[tool.poetry.dependencies]
# Acepta 3.10 a 3.13 (tú estás en 3.13)
python = ">=3.10,<3.14"

fastapi = ">=0.110,<1.0"
uvicorn = { extras = ["standard"], version = ">=0.37.0,<0.38.0" }
python-dotenv = "^1.0"
supabase = "^2.6"
httpx = "^0.27"
psycopg = { extras = ["binary"], version = ">=3.2.10,<4.0.0" }
pydantic = { extras = ["email"], version = ">=2.12.3,<3.0.0" }
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
